* main: (39 commits)
  Define USE_SZIP variable for nc-config.cmake.in
  matching cmake variables in autotools configuration
  moving the version into the project command in cmake
  Updated doxygen files for older, less-forgiving versions of doxygen (1.9.1, at least)
  Clean up doxygen warnings that were being treated as failures.
  Correcting a weird doxygen issue that has appeared.
  cmake: Fix Szip link using correct cmake var
  cmake: Improve FindSzip logic to provide a Szip_LIBRARY var
  Changed link to netCDF-Fortran documentation.
  Replaced ancient K&R function declarations to be C23 compatible
  Rebased PR by hand against main.
  count argument in H5Sselect_hyperslab
  Add H5FD_http_finalize function and call on hdf5 finalize
  Catching up on PRs, this is 2431 on the current 'main'
  Fix cmake syntax typo.
  Typo fix in support of gh2824.wif
  Fix typo in nc-config.cmake.in
  Removed a use of sprintf that required changing a function signature
  Replaced some sprintf with snprintf with aid of new variable containing size
  Replaced trivial uses of sprintf with snprintf
  ...

+++ include/nc4internal.h
@@ -131,7 +131,7 @@ typedef struct NC_OBJ
 {
     NC_SORT sort; /**< Type of object. */
     char* name;   /**< Name, assumed to be null terminated. */
-    size_t id;    /**< This objects ID. */
+    int id;    /**< This objects ID. */
 } NC_OBJ;
 
 /**
